Scaffold LSD Depository

+++ programs/uxd/src/instructions/initialize_identity_depository.rs
@@ -100,9 +100,10 @@ pub(crate) fn handler(ctx: Context<InitializeIdentityDepository>) -> Result<()> 
     depository.collateral_vault = ctx.accounts.collateral_vault.key();
     depository.collateral_vault_bump = collateral_vault_bump;
     depository.redeemable_amount_under_management = u128::MIN;
-    depository.redeemable_amount_under_management_cap = DEFAULT_REDEEMABLE_UNDER_MANAGEMENT_CAP
-        .checked_mul(redeemable_mint_unit)
-        .ok_or_else(|| error!(UxdError::MathError))?;
+    depository.redeemable_amount_under_management_cap =
+        u128::from(DEFAULT_REDEEMABLE_UNDER_MANAGEMENT_CAP)
+            .checked_mul(redeemable_mint_unit)
+            .ok_or_else(|| error!(UxdError::MathError))?;
     depository.minting_disabled = true;
 
     depository.mango_collateral_reinjected_wsol = false;
